[TASK] Convert some configuration check tests from functional to unit
Also fix the name of a testcase file.

Part of #928
