+++ Tests/Unit/Configuration/MyVipEventsConfigurationCheckTest.php
@@ -19,9 +19,16 @@ final class MyVipEventsConfigurationCheckTest extends UnitTestCase
      */
     private $subject;
 
+    /**
+     * @var DummyConfiguration
+     */
+    private $configuration;
+
     protected function setUp(): void
     {
-        $this->subject = new MyVipEventsConfigurationCheck(new DummyConfiguration(), 'plugin.tx_seminars_pi1');
+        $this->configuration = new DummyConfiguration();
+
+        $this->subject = new MyVipEventsConfigurationCheck($this->configuration, 'plugin.tx_seminars_pi1');
     }
 
     /**
@@ -30,5 +37,41 @@ final class MyVipEventsConfigurationCheckTest extends UnitTestCase
     public function isConfigurationCheck(): void
     {
         self::assertInstanceOf(AbstractConfigurationCheck::class, $this->subject);
+    }
+
+    /**
+     * @test
+     */
+    public function checkWithEmptyConfigurationCreatesErrors(): void
+    {
+        $this->subject->check();
+
+        $result = $this->subject->getWarningsAsHtml();
+        self::assertNotSame([], $result);
+    }
+
+    /**
+     * @test
+     */
+    public function checkWithEmptyConfigurationUsesProvidedNamespaceForErrors(): void
+    {
+        $this->subject->check();
+
+        $result = $this->subject->getWarningsAsHtml();
+        self::assertArrayHasKey(1, $result);
+        self::assertStringContainsString('plugin.tx_seminars_pi1', $result[1]);
+    }
+
+    /**
+     * @test
+     */
+    public function checkWithEmptyConfigurationWithManagerPermissionsEnabledUsesProvidedNamespaceForErrors(): void
+    {
+        $this->configuration->setAsBoolean('mayManagersEditTheirEvents', true);
+
+        $this->subject->check();
+
+        $result = $this->subject->getWarningsAsHtml();
+        self::assertNotSame([], $result);
     }
 }
